[FEAT] #3 - refactored PostsVC, VM, View
