+++ EasyVel/Presentation/Posts/KeywordPostsVCFactory.swift
@@ -15,7 +15,7 @@ final class KeywordPostsVCFactory {
         tag: String
     ) -> PostsViewController {
         let vc = PostsViewController(
-            viewModel: PostsViewModel(viewType: .keyword, tag: tag, service: DefaultPostService.shared)
+            viewModel: PostsViewModel(viewType: .keyword, tag: tag)
         )
         return vc
     }
@@ -28,9 +28,8 @@ final class KeywordPostsVCFactory {
         postDTOList: [PostDTO]
     ) -> PostsViewController {
         let vc = PostsViewController(
-            viewModel: PostsViewModel(viewType: .keyword, tag: tag,service: DefaultPostService.shared),
-            isNavigationBarHidden: isNavigationBarHidden,
-            posts: postDTOList
+            viewModel: PostsViewModel(viewType: .keyword, tag: tag),
+            posts: postDTOList, isNavigationBarHidden: isNavigationBarHidden
         )
         return vc
     }
